services/data-pipeline/requirements.txt
# Data Pipeline Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.5.2
httpx==0.25.2
python-dotenv==1.0.0

# Vector Database and Embeddings
chromadb==0.4.22
sentence-transformers>=2.7.0
numpy==1.24.3

# LLM Integration
google-generativeai==0.3.2
tiktoken>=0.5.0

# Data Processing
markdown==3.5.1
beautifulsoup4==4.12.2
python-multipart==0.0.6

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# OpenTelemetry and Observability (Phase 1 - LangSmith Integration)
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation==0.45b0

# Sentry removed for simplified setup

# System monitoring
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
